[Week9][Improvement] Refactor Main Activity and Character View Model
* [X] Move periodic worker code from Main Activity to Character View Model
